+++ src/Git_Updater/Traits/GU_Trait.php
@@ -87,6 +87,26 @@ trait GU_Trait {
 	}
 
 	/**
+	 * Check if we should run on the current page.
+	 *
+	 * @return bool
+	 */
+	final public static function should_run_on_current_page(): bool {
+		global $pagenow;
+
+		$pages            = [ 'update-core.php', 'update.php', 'plugins.php', 'themes.php' ];
+		$view_details     = [ 'plugin-install.php', 'theme-install.php' ];
+		$autoupdate_pages = [ 'admin-ajax.php', 'index.php', 'wp-cron.php' ];
+		$settings_pages   = is_multisite() ? [ 'settings.php', 'edit.php' ] : [ 'options.php', 'options-general.php' ];
+
+		if ( ! in_array( $pagenow, array_merge( $pages, $view_details, $autoupdate_pages, $settings_pages ), true ) ) {
+			return false;
+		}
+
+		return true;
+	}
+
+	/**
 	 * Get cache key.
 	 *
 	 * @param  string|bool $repo Repo name or false.