+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * fix `flush-repo-cache` REST endpoint, was getting caught in `$existing_cache`
+* added `should_run_on_current_page()` and check pages for loading certain parts of plugin
 
 #### 12.24.0 / 2026-03-11
 * update erusev/parsedown to 1.8